+++ idf_build_apps/app.py
@@ -107,6 +107,9 @@ class App(object):
         self.parallel_index = 1
         self.parallel_count = 1
 
+        self._collect_app_info = None
+        self._collect_size_info = None
+
         # sdkconfig attrs, use properties instead
         self._sdkconfig_defaults = self._get_sdkconfig_defaults(sdkconfig_defaults_str)
         self._sdkconfig_files = None
@@ -243,6 +246,20 @@ class App(object):
     def size_json_path(self):
         if self._size_json_path:
             return os.path.join(self.build_path, self._expand(self._size_json_path))
+
+        return None
+
+    @property
+    def collect_app_info(self):
+        if self._collect_app_info:
+            return self._expand(self._collect_app_info)
+
+        return None
+
+    @property
+    def collect_size_info(self):
+        if self._collect_size_info:
+            return self._expand(self._collect_size_info)
 
         return None
 
@@ -385,19 +402,6 @@ class App(object):
             check=True,
         )
         LOGGER.info('=> Generated size info to %s', self.size_json_path)
-
-    def collect_size_info(self, output_fs):  # type: (t.TextIO) -> None
-        if not os.path.isfile(self.size_json_path):
-            self.write_size_json()
-
-        size_info_dict = {
-            'app_name': self.name,
-            'config_name': self.config_name,
-            'target': self.target,
-            'path': self.size_json_path,
-        }
-        output_fs.write(json.dumps(size_info_dict) + '\n')
-        LOGGER.info('=> Recorded size info file path in %s', output_fs.name)
 
     def to_json(self):
         # keeping backward compatibility, only provide these stuffs